Fix bug with reservation during sermon being done twice
Somehow reservation of the altar and the cell which the preacher stands on, is released before the the FinishAction of the JobDriver releases it. Since the reservation lasts right up until the moment when the sermon finishes, finding the second way the reservation is being released is unneccessary. It is a lot easier to remove the releasing in the FinishAction instead.